[no release notes] move CassandraClientPoolImpl's metrics out into their own class to simplify it a bit.  I've kept the per host metrics code even though this is disabled right now and will never do anything as it explains some of the complexity and makes it easier to reenable
